requirements.txt
paho-mqtt==1.6.1
matplotlib==3.5.1
python-dotenv==0.19.2